update: queries structures
