+++ src/graphql/queries/consensus.ts
@@ -1,59 +1,8 @@
 import { gql } from '@apollo/client';
 
-export const LATEST_BLOCK_HEIGHT_QUERY = gql`
-  query LatestBlockHeight {
-    block(limit: 1, order_by: {height: desc}) {
-      height
-  }
-}`;
-
-export const AVERAGE_BLOCK_TIME_FROM_GENESIS_QUERY = gql`
-query AverageBlockTimeAllTime{
-  average_block_time_from_genesis(limit: 1, order_by: {height: desc}){
-    average_time
-  }
-}`;
-
-export const AVERAGE_BLOCK_LAST_MINUTE_QUERY = gql`
-query AverageBlockTimeLastMinute{
-  average_block_time_per_minute(limit: 1, order_by: {height: desc}){
-    average_time
-  }
-}`;
-
-export const AVERAGE_BLOCK_LAST_HOUR_QUERY = gql`
-query AverageBlockTimeLastHour{
-  average_block_time_per_hour(limit: 1, order_by: {height: desc}){
-    average_time
-  }
-}`;
-
-export const AVERAGE_BLOCK_LAST_DAY_QUERY = gql`
-query AverageBlockTimeLastDay{
-  average_block_time_per_day(limit: 1, order_by: {height: desc}){
-    average_time
-  }
-}`;
-
-// blockInfoFromLimit return the last n block info
-export const BLOCK_INFO_LIMIT = gql`
-subscription LatestBlocks($limit: Int!) {
-  block(limit: $limit, order_by: {height: desc}) {
-    hash
-    proposer {
-      validator_infos {
-        validator_descriptions(order_by: {timestamp: desc}) {
-          moniker
-          operator_address
-        }
-      }
-    }
-    height
-    num_txs
-    timestamp
-  }
-}`;
-
+/**
+ * Get the consensus state
+ */
 export const CONSENSUS_STATE = gql`
   subscription consensus_state{
     consensus{
