+++ src/graphql/queries/proposals.ts
